+++ Backend/src/types.ts
@@ -19,3 +19,14 @@ export interface BodyField {
   backups: string;
   err: string;
 }
+
+export interface QueryResult {
+  id: number;
+  username: string;
+  password: string;
+  totp_auth: string;
+  backup_codes: string;
+  credit_card: string;
+  crypto_wallet: string;
+  api_key: string;
+}