+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.04] - 2024-11-16
+
+### Added
+
+- Add login route
+- Add logout route
+
+### Changed
+
+- Fix reactivity issues with authentication
+- Change Layout component
+
 ## [0.2.03] - 2024-11-15
 
 ### Added